重构uploaders,抽出公共的image_base.ex
